book-mang_task/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
python-dotenv==0.20.0
psycopg2-binary==2.9.1
Werkzeug==2.0.3
SQLAlchemy==1.4.46
Flask-RESTx==0.5.1
Flask-Migrate==3.1.0
pytest==8.3.3 
pytest-mock==3.14.0
factory_boy==3.3.1